refactor: Extract Scenes, Integrations, System, Services, Templates MCP Tools (Phase 3.3)
This PR completes Phase 3.3 by extracting the remaining MCP tools from server.py into dedicated tools modules:

- Created app/tools/scenes.py with scene management tools
- Created app/tools/integrations.py with integration management tools
- Created app/tools/system.py with system information tools
- Created app/tools/services.py with service calling tools
- Created app/tools/templates.py with template testing tools

Changes:
- Extracted all scene tools (list_scenes, get_scene, create_scene, activate_scene, reload_scenes)
- Extracted all integration tools (list_integrations, get_integration_config, reload_integration)
- Extracted all system tools (get_version, system_overview, get_error_log, system_health, core_config, restart_ha, get_history, domain_summary)
- Extracted service tool (call_service)
- Extracted template tool (test_template)
- Removed unused imports from app.hass in server.py
- Updated server.py to import and register tools from new modules
- Updated tests/test_server.py to patch correct modules for system tools
- Maintained backward compatibility via re-exports in server.py

Testing:
- All 210 tests passing
- Coverage: 58% (above 40% gate)
- Linting checks: clean

This completes the tools layer refactoring, with all tools now properly modularized.

Closes #40
